cloudbuildsubscriber.Receive takes in pubsub message as argument
